Refactored format db to move some common code to the parent
